Compiler introspect (#100)
* compiler info in own python module

* clang version is gotten with a precision of 2

* debug

* version

* version

* version

* remove debug messages

* close file after running ConanBuildInfoParser

* CONAN_COMPILER_VERSION env variable is optional in cibuildwheel

* remove output_library_name

* get_conan_options

* SONAR_SCANNER_VERSION=4.7.0.2747

* whitespace

* find_missing_libraries has strategies

* add libs and compiler metadata to extension

* remove test conan executable from builder

* reuse conan build with cibuildwheel

* debug

* pass CONAN_COMPILER_LIBCXX to cibuildwheel

* get_gcc_version checks version

* get_gcc_version checks version

* debug

* debug

* get_gcc_version

* get_gcc_version

* debug

* debug

* debug

* debug

* compiler version

* debug

* debug

* debug

* debug

* debug

* ubuntu gcc 10.2

* version information for the compilers in github actions

* verbosity on tox matrix up to 3

* libstdc++11 to ci

* libstdc++11 to ci

* libstdc++ to ci on macs

* libc++ to ci on macs

* libc++ to ci on macs

* conan defaults for libstdc++ on linux tox

* use Pybind11Extension

* add libraries in the correct order

* on linux use shared library for tiff

* on linux use shared library for tiff

* only copy files in lib dir from conan if exists

* skip testing of tesseract on linux

* handel inplace

* handel inplace

* linux only shared lib tiff

* linux only shared lib iconv

* update compiler version

* import libtiffxx.so

* add library build path to extension library path

* import conan imports directly to lib dir

* debug missing conanbuildinfo.txt file

* toml

* handle inplace builds

* handle inplace builds runtime_library_dirs

* debug message

* debug message

* debug message

* debug message

* debug message

* debug message

* use locate_conanbuildinfo

* use locate_conanbuildinfo

* use locate_conanbuildinfo

* use locate_conanbuildinfo

* conanfileinfo_locations added release

* debug

* dll tester includes same dir as the file

* include path

* debug

* look locally

* debug

* debug

* zlib explicit

* use static files again

* compiler version updated

* docker install conan in base python

* set compiler version

* remove extra env

* debug

* dot notation

* dot notation

* wheel~=0.37

* wheel~=0.37

* wheel==0.37

* use gcc 10 in packages

+++ setup.py
@@ -1,5 +1,7 @@
 import os
 import sys
+from distutils import ccompiler
+from pathlib import Path
 
 import setuptools
 import shutil
@@ -7,36 +9,93 @@ from distutils.version import StrictVersion
 
 sys.path.insert(0, os.path.dirname(__file__))
 from builders.deps import get_win_deps
-
+try:
+    from pybind11.setup_helpers import Pybind11Extension
+except ImportError:
+    from setuptools import Extension as Pybind11Extension
 cmd_class = {}
 try:
-    from builders.conan_libs import BuildConan
-    cmd_class["build_conan"] = BuildConan
+    from builders import conan_libs
+    cmd_class["build_conan"] = conan_libs.BuildConan
 except ImportError:
     pass
 
 try:
-    from builders.pybind11_builder import BuildPybind11Extension
+    from builders.pybind11_builder import BuildPybind11Extension, UseSetuptoolsCompilerFileLibrary
+
+
+    def test_tesseract(build_file: str):
+        with open(build_file, "r") as f:
+            parser = conan_libs.ConanBuildInfoParser(f)
+            data = parser.parse()
+        path = data['bindirs_tesseract']
+        tesseract = shutil.which("tesseract", path=path[0])
+
+        tester = {
+            'darwin': conan_libs.MacResultTester,
+            'linux': conan_libs.LinuxResultTester,
+            'win32': conan_libs.WindowsResultTester
+        }.get(sys.platform)
+
+        if tester is None:
+            raise AttributeError(f"unable to test for platform {sys.platform}")
+
+        compiler = ccompiler.new_compiler()
+        tester = tester(compiler)
+        libs_dirs = data['libdirs']
+        for libs_dir in libs_dirs:
+            tester.test_shared_libs(libs_dir)
+        tester.test_binary_dependents(Path(tesseract))
+
 
     class BuildTesseractExt(BuildPybind11Extension):
 
-        def build_extension(self, ext):
-            missing = self.find_missing_libraries(ext)
-
+        def build_extension(self, ext: Pybind11Extension):
+            missing = self.find_missing_libraries(ext, strategies=[
+                UseSetuptoolsCompilerFileLibrary(
+                    compiler=self.compiler,
+                    dirs=self.library_dirs + ext.library_dirs
+                ),
+            ])
+            build_conan = self.get_finalized_command("build_conan")
             if len(missing) > 0:
                 self.announce(f"missing required deps [{', '.join(missing)}]. "
                               f"Trying to get them with conan", 5)
-                self.run_command("build_conan")
+                build_conan.build_libs = ['outdated']
+            else:
+                build_conan.build_libs = []
+            build_conan.run()
+            # This test os needed because the conan version keeps silently
+            # breaking the linking to openjp2 library.
+            conanfileinfo_locations = [
+                self.build_temp,
+                os.path.join(self.build_temp, "Release"),
+            ]
+            conan_info_dir = os.environ.get('CONAN_BUILD_INFO_DIR')
+            if conan_info_dir is not None:
+                conanfileinfo_locations.insert(0, conan_info_dir)
+            for location in conanfileinfo_locations:
+                conanbuildinfo = os.path.join(location, "conanbuildinfo.txt")
+                if os.path.exists(conanbuildinfo):
+                    test_tesseract(conanbuildinfo)
+                    break
             super().build_extension(ext)
+            tester = {
+                'darwin': conan_libs.MacResultTester,
+                'linux': conan_libs.LinuxResultTester,
+                'win32': conan_libs.WindowsResultTester
+            }.get(sys.platform)
+            dll_name = \
+                os.path.join(self.build_lib, self.get_ext_filename(ext.name))
+
+            tester().test_binary_dependents(Path(dll_name))
 
         def run(self):
-            pybind11_include_path = self.get_pybind11_include_path()
-
-            if pybind11_include_path is None:
-                raise FileNotFoundError("Missing pybind11 include path")
-
-            self.include_dirs.append(pybind11_include_path)
             super().run()
+            def locate(dep, location):
+                for f in os.scandir(dest):
+                    if f.name.lower() == dep.lower():
+                        return f.path
 
             for e in self.extensions:
 
@@ -51,9 +110,15 @@ try:
                     dest = os.path.dirname(dll_name)
 
                     for dep in deps:
-                        dll = self.find_deps(dep)
+                        if os.path.exists(os.path.join(dest, dep)):
+                            print(f"Package already has {dep}")
+                            continue
+                        paths = os.environ['path'].split(";")
+                        dll = self.find_deps(dep, paths)
                         if dll is None:
-                            raise FileNotFoundError("Missing for {}".format(dep))
+                            for i in os.scandir(dest):
+                                print(i.path)
+                            raise FileNotFoundError(f"Missing {dep}. Searched {paths}")
                         shutil.copy(dll, dest)
 
     cmd_class["build_ext"] = BuildTesseractExt
@@ -69,8 +134,8 @@ if StrictVersion(setuptools.__version__) < StrictVersion('30.3'):
           )
 
     sys.exit(1)
-
-tesseract_extension = setuptools.Extension(
+# tesseract_extension = setuptools.Extension(
+tesseract_extension = Pybind11Extension(
     "uiucprescon.ocr.tesseractwrap",
     sources=[
         'uiucprescon/ocr/Capabilities.cpp',
